requirements.txt
flask==2.0.3
werkzeug==2.0.3
gradio==3.25.0
scikit-learn==1.2.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.28.1
torch==2.0.0
textblob==0.17.1
symspellpy==6.7.7
pandas==2.0.0
numpy==1.24.2
python-dotenv==1.0.0
requests==2.28.2 